+++ src/channel.hh
@@ -81,7 +81,7 @@ class Channel {
   Channel(const Options& options, int which_channel, std::ostream& output_stream);
   Channel(const Options& options, std::ostream& output_stream, std::uint16_t pi);
   void processBit(bool bit, std::size_t which_stream);
-  void processBits(const BitBuffer& buffer, std::size_t which_stream);
+  void processBits(const BitBuffer& buffer);
   void processGroup(Group group, std::size_t which_stream);
   void flush();
   float getSecondsSinceCarrierLost() const;
@@ -96,6 +96,8 @@ class Channel {
   Station station_;
   RunningAverage<float, kNumBlerAverageGroups> bler_average_;
   std::chrono::time_point<std::chrono::system_clock> last_group_rx_time_;
+  // When the first bit of this group was received
+  std::array<DelayLine<double, 104 + 1>, 4> delayed_time_offset_;
 };
 
 }  // namespace redsea