+++ src/io/output.hh
@@ -1,6 +1,7 @@
 #ifndef REDSEA_OUTPUT_HH
 #define REDSEA_OUTPUT_HH
 
+#include <iomanip>
 #include <ostream>
 
 #include "src/groups.hh"
@@ -11,11 +12,16 @@ namespace redsea {
 
 inline void printAsHex(const Group& group, const Options& options, std::ostream& output_stream) {
   if (!group.isEmpty()) {
-    if (group.getDataStream() > 0)
+    if (group.getDataStream() > 0) {
       output_stream << "#S" << group.getDataStream() << " ";
+    }
     group.printHex(output_stream);
-    if (options.timestamp)
+    if (options.timestamp) {
       output_stream << ' ' << getTimePointString(group.getRxTime(), options.time_format);
+    }
+    if (options.time_from_start && group.getTimeFromStart().valid) {
+      output_stream << ' ' << std::fixed << std::setprecision(6) << group.getTimeFromStart().data;
+    }
     output_stream << '\n' << std::flush;
   }
 }